# openapi.yaml
openapi: 3.1.0
info:
  title: INLABS DOU Middleware
  version: 1.0.0
servers:
  - url: https://SEU-DOMINIO-RAILWAY-AQUI
paths:
  /dou/transmissao:
    post:
      operationId: buscar_dou_transmissao
      summary: Autentica no INLABS (sessão limpa + login forçado) e retorna publicações relevantes p/ transmissão
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  description: "Data no formato YYYY-MM-DD. Se omitido, usar DOU do dia; se não houver, última edição."
                include_pdf:
                  type: boolean
                  description: "Se true, o middleware pode retornar PDF em base64 (quando implementado)."
              additionalProperties: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  date_used: { type: string }
                  edition_info: { type: string }
                  source: { type: string, enum: [INLABS] }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        section: { type: string }
                        organ: { type: string }
                        act_type: { type: string }
                        title: { type: string }
                        url: { type: string }
                        full_text: { type: string }
                        triggers: { type: array, items: { type: string } }
                        entities_matched: { type: array, items: { type: string } }
                  pdf:
                    type: object
                    properties:
                      filename: { type: string }
                      content_base64: { type: string }
                required: [date_used, items]
        "401":
          description: Unauthorized
        "409":
          description: Captcha/MFA detectado
        "500":
          description: Erro interno
components:
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - api_key: []